Redo logging, add constants for paths
Reworked logging to save a frame every 5 minutes for
comms. Also eliminated saving of a video in favor
of saving a frame every 5 minutes. Eliminated iteration
of j within log_frames() as a global. Added constants for
paths, and will save logs and frames for each run in
separate subdirectories created in main.